[sw/device] Second stage ROM
This extracts the second stage ROM from the current ROM code.

The second stage ROM is patched and loaded by the base ROM, and
implements the reminder of the ROM boot sequence, up to loading and
booting the first mutable firmware image from the external flash
controller.

Although this change does not implement the full second stage ROM set of
functionalities, it starts after being loaded by the base ROM.

Signed-off-by: Samuel Ortiz <sameo@rivosinc.com>
Signed-off-by: Loïc Lefort <loic@rivosinc.com>