[silicon_creator/lib] auto-load attestation keygen seed from flash info
The current OTBN boot services API required the user to input the
additional DRBG seed manually. Since this seed is provisioned into a
flash info page at manufacturing time, this updates the code to
automatically load the specified seed from the flash instead. This
simplifies the user experience of driving this library.

Signed-off-by: Tim Trippel <ttrippel@google.com>