[silicon_creator] autoformat attestation key seed flash page
This updates OTBN boot services lib to erase the attestation key seed
flash info page if it detects it has not been provisioned (i.e.,
an integrity error is encountered during a read). Additionally a warning
is printed over the console if this happens.

This enables ROM_EXT E2E tests to pass without having to pre-configure
flash info pages as if a chip was already provisioned.

Signed-off-by: Tim Trippel <ttrippel@google.com>