[pentest] Move SHA3 trigger arm&disarm
As discussed offline with @vogelpi, this PR moves the arming and
disarming of the SCA trigger for SHA3 before and after the
process command is issued. This ensures that we only capture the
SCA relevant parts for doing power analysis.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 1a91b7e07985716275be8d95508e850ddaae14fe)