[pentest] Correctly set unused key shares
This commit fixes a bug in the simpleserial and uJSON AES SCA
application. Inside the aes_key_mask_and_config function,
two for loops are responsible for setting the key shares. The
first for loop sets all key shares up to the provided key_length.
The second for loop sets all unused key bits.

Previously, the second for loop did not set the key shares for the
unused bits correctly. For example, when key_len=16, only key shares
0...3 were correctly set, key shares 4...7 were uninitialized, e.g.
contained old values from memory. However, for SCA we assume that
the unused key share values are set to 0.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>