[aes/sw] Load all-zero vector into PRNG buffer stage for masking off
The input masks are now taken from the PRNG buffer stage. This means
after loading the magic seed into the PRNG and having the PRNG output
the all-zero vector, the buffer stage needs to be updated as well.
This can be achieved by triggering via a clearing operation of the data
output registers.

This resolves lowRISC/OpenTitan#22917.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>